+++ PactPort/DynamicObjectAccess/Evaluators/NullEvaluator.cs
@@ -1,0 +1,20 @@
+namespace Pact.Provider.Wrapper.PactPort.DynamicObjectAccess.Evaluators
+{
+    public class NullEvaluator:IEvaluator
+    {
+        public bool IsFlatEnough(object data)
+        {
+            return true;
+        }
+
+        public object Evaluate(object data)
+        {
+            return null;
+        }
+
+        public bool Supports(object data)
+        {
+            return false;
+        }
+    }
+}