+++ PactPort/DynamicObjectAccess/Evaluators/PrimitiveEvaluator.cs
@@ -1,0 +1,31 @@
+namespace Pact.Provider.Wrapper.PactPort.DynamicObjectAccess.Evaluators
+{
+    public class PrimitiveEvaluator:IEvaluator
+    {
+        public bool IsFlatEnough(object data)
+        {
+            return IsPrimitive(data);
+        }
+
+        public object Evaluate(object data)
+        {
+            return data;
+        }
+
+        public bool Supports(object data)
+        {
+            return true;
+        }
+
+
+        private bool IsPrimitive(object data)
+        {
+            if (data == null)
+            {
+                return false;
+            }
+
+            return data is string || data.GetType().IsPrimitive;
+        }
+    }
+}